refactor(confluence-connector): move external-id logic to dedicated file
Extract parseExternalId, buildExternalId, and ParsedExternalId from
ingestion.constants.ts into constants/external-id.ts. These are
scope-related, not ingestion-related.

UN-18296
