+++ services/confluence-connector/src/constants/ingestion.constants.ts
@@ -11,8 +11,6 @@ export type EnabledDisabledMode = (typeof EnabledDisabledMode)[keyof typeof Enab
 
 export type IngestionMode = (typeof IngestionMode)[keyof typeof IngestionMode];
 
-export const EXTERNAL_ID_PREFIX = 'confc:' as const;
-
 export const IngestionSourceKind = {
   Cloud: 'ATLASSIAN_CONFLUENCE_CLOUD',
   DataCenter: 'ATLASSIAN_CONFLUENCE_ONPREM',
@@ -22,36 +20,6 @@ export type IngestionSourceKind = (typeof IngestionSourceKind)[keyof typeof Inge
 
 export function getSourceKind(instanceType: 'cloud' | 'data-center'): IngestionSourceKind {
   return instanceType === 'cloud' ? IngestionSourceKind.Cloud : IngestionSourceKind.DataCenter;
-}
-
-export interface ParsedExternalId {
-  tenantName: string;
-  spaceId: string;
-  spaceKey: string;
-}
-
-export function parseExternalId(externalId: string | undefined): ParsedExternalId | undefined {
-  if (!externalId) {
-    return undefined;
-  }
-
-  if (!externalId.startsWith(EXTERNAL_ID_PREFIX)) {
-    return undefined;
-  }
-
-  const withoutPrefix = externalId.slice(EXTERNAL_ID_PREFIX.length);
-  const segments = withoutPrefix.split(':');
-
-  if (segments.length !== 3) {
-    return undefined;
-  }
-
-  const [tenantName, spaceId, spaceKey] = segments as [string, string, string];
-  return { tenantName, spaceId, spaceKey };
-}
-
-export function buildExternalId(tenantName: string, spaceId: string, spaceKey: string): string {
-  return `${EXTERNAL_ID_PREFIX}${tenantName}:${spaceId}:${spaceKey}`;
 }
 
 export const INGESTION_MIME_TYPE = 'text/html' as const;
